The final release of alfred
