+++ src/main/java/alfred/task/TaskList.java
@@ -1,4 +1,4 @@
-package duke.task;
+package alfred.task;
 
 import java.util.ArrayList;
 import java.util.stream.Collectors;
