+++ text-ui-test/input.txt
@@ -1,3 +1,3 @@
-duke.task.Todo borrow book
-duke.task.Deadline return book /by Sunday
-duke.task.Event project meeting /from Mon 2pm /to 4pm
+alfred.task.Todo borrow book
+alfred.task.Deadline return book /by Sunday
+alfred.task.Event project meeting /from Mon 2pm /to 4pm
